Fix disabled merge button on conference failure.
There is currently no direct plumbing of the conference failure
to InCallUI. Instead it depends on the capabilities refresh.
In the conference failure scenario, there is no such change
and as a result the merge button remains greyed out until a
subsequent update using state or extras. The fix here is to
create a dummy extra to refresh the UI.

Change-Id: Id487418aa4729f822ba0a19661805e1bf3ccbb6a
CRs-Fixed: 931958